+++ src/graphql_clj/validator/rules/variables_are_input_types.clj
@@ -2,16 +2,16 @@
   "A GraphQL operation is only valid if all the variables it defines are of input types (scalar, enum, or input object)."
   (:require [graphql-clj.visitor :refer [defnodevisitor]]
             [graphql-clj.validator.errors :as ve]
-            [clojure.spec :as s]))
+            [graphql-clj.spec :as spec]))
 
-(defn- bad-variable-type-error [{:keys [spec argument-name value]}]
-  (let [type-name (name (s/get-spec spec))]
-    (format "Argument '%s' of type '%s' has invalid value: %s. Reason: %s."
-            argument-name type-name (ve/render value) (ve/explain-invalid spec value))))
+(defn- bad-variable-type-error [{:keys [variable-name]} {:keys [spec]}]
+  (format "Variable '$%s' cannot be non-input type '%s'." variable-name (name spec)))
 
-(defnodevisitor bad-variable-type :pre :variable-definition
-  [{:keys [spec value v/path] :as n} s]
-  (when (and spec value (not (ve/valid? spec value path)))
-    {:state (ve/update-errors s (bad-variable-type-error (assoc n :value value)))}))
+(def acceptable-types #{:scalar :enum-definition :input-definition})
+
+(defnodevisitor bad-variable-type :pre :variable-definition [n s]
+  (let [{:keys [node-type] :as type-node} (spec/get-type-node n s)]
+    (when-not (acceptable-types node-type)
+      {:state (ve/update-errors s (bad-variable-type-error n type-node))})))
 
 (def rules [bad-variable-type])