+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file. This change
 - Build a map from namespaced spec keyword to their corresponding nodes
 - Include full parent node in each child node (for getting to parent types)
 - Validate argument, type, and fragment names
+- Validate that variables are input types
 
 ## [0.1.16] - 2016-10-12
 ### BREAKING